+++ example/android/app/src/main/kotlin/tech/graaf/franz/ar_flutter_plugin_example/MainActivity.kt
@@ -1,0 +1,6 @@
+package tech.graaf.franz.ar_flutter_plugin_example
+
+import io.flutter.embedding.android.FlutterActivity
+
+class MainActivity: FlutterActivity() {
+}