+++ example/lib/gloabl_variables.dart
@@ -1,0 +1,6 @@
+class GlobalVariables {
+  // static String arObjectUrl1 =
+  //     "https://raw.githubusercontent.com/KhronosGroup/glTF-Sample-Models/main/2.0/Duck/glTF-Binary/Duck.glb";
+  static String arObjectUrl1 =
+      "https://tranvubao-dev.github.io/ar_flutter_plugin/girl3.glb";
+}