+++ ios/Classes/SwiftArFlutterPluginPlus.swift
@@ -1,0 +1,18 @@
+import Flutter
+import UIKit
+
+public class SwiftArFlutterPluginPlus: NSObject, FlutterPlugin {
+  public static func register(with registrar: FlutterPluginRegistrar) {
+    let channel = FlutterMethodChannel(name: "ar_flutter_plugin_plus", binaryMessenger: registrar.messenger())
+    let instance = SwiftArFlutterPluginPlus()
+    registrar.addMethodCallDelegate(instance, channel: channel)
+    
+    let factory = IosARViewFactory(messenger: registrar.messenger())
+    registrar.register(factory, withId: "ar_flutter_plugin_plus")
+  }
+
+  public func handle(_ call: FlutterMethodCall, result: @escaping FlutterResult) {
+    result("iOS " + UIDevice.current.systemVersion)
+  }
+
+}