+++ lib/awesome_utils.dart
@@ -1,9 +1,0 @@
-/// A Calculator.
-// class Calculator {
-//   /// Returns [value] plus 1.
-//   int addOne(int value) => value + 1;
-// }
-
-library awesome_utils;
-
-int add(int a, int b) => a + b;