+++ test/ar_flutter_plugin_test.dart
@@ -1,0 +1,23 @@
+import 'package:ar_flutter_plugin_plus/ar_flutter_plugin_plus.dart';
+import 'package:flutter/services.dart';
+import 'package:flutter_test/flutter_test.dart';
+
+void main() {
+  const MethodChannel channel = MethodChannel('ar_flutter_plugin_plus');
+
+  TestWidgetsFlutterBinding.ensureInitialized();
+
+  setUp(() {
+    channel.setMockMethodCallHandler((MethodCall methodCall) async {
+      return '42';
+    });
+  });
+
+  tearDown(() {
+    channel.setMockMethodCallHandler(null);
+  });
+
+  test('getPlatformVersion', () async {
+    expect(await ArFlutterPluginPlus.platformVersion, '42');
+  });
+}