+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+# Changelog
+
 ## 0.0.1
 
-* TODO: Describe initial release.
+- Extended the original ar_flutter_plugin with image tracking.
+
+## 1.0.1
+- Package rename to ar_flutter_holi
+- Plugin structure updated
+- Build artifacts cleaned
+- Documentation updates